+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 CHANGELOG
 =========
 
+1.6.0
+-----
+
+Added functionality:
+
+ * Added LetterCase transformer
+ * Added JsonEncode, JsonDecode transformer
+ * Upgrade phpunit to version 5.5
+ 
+
 1.5.0
 -----